# modulo-8/openapi/receipts.openapi.yaml
openapi: 3.0.3
info:
  title: M8 - Notificaciones, Documentos y Soporte (Comprobantes PDF)
  description: |
    Microservicio de la **Plataforma Distribuida de Movilidad Urbana (ISI - Paradigmas 3)**.
    Responsable de la **emisión, consulta, descarga y reenvío de comprobantes de viaje en PDF**.

    **Versión 2.3.0: evolución individual AE2 (Juan Gualtieri) sobre la versión 1.0.0 del Grupo 14 (AE1).**
    - **RF-8.3:** Comprobante PDF. Además de `POST /receipts`, se emite de forma asincrónica al consumir
      `payment.confirmed` (RabbitMQ). Persistencia en PostgreSQL (esquema `receipts`).
    - **RF-8.4:** Reenvío de comprobante (entrega simulada) y descarga mediante enlace temporal (Redis).
    - **Idempotencia (RNF-08/09):** un único comprobante por `tripId`, garantizado por una restricción
      `UNIQUE` de la base aun con varias réplicas del servicio.
    - **Autorización fiscal (2.1.0):** antes de emitir, el comprobante se autoriza ante un servicio externo
      (simulado), con timeout de 2 s y circuit breaker. El código queda en `fiscal` e impreso en el PDF.
    - **Pago verificado en M7 (2.2.0):** antes de emitir se consulta `GET /metodo-pago/{viajeId}` de M7.
      Solo un pago `autorizado` genera comprobante, y el medio de pago se toma de M7.
    - **Importe de M7 (2.3.0):** el total y la moneda del comprobante se toman de M7 cuando los informa.
    - **Dependencias caídas:** responden `503` con `Retry-After` (nunca `500`); el servicio sigue atendiendo.

...
servers:
...
security: []
...
paths:
  /receipts:
    post:
      tags:
        - Comprobantes (RF-8.3 / RF-8.4)
      summary: Emite el comprobante PDF de un viaje finalizado (RF-8.3)
      description: |
        Recibe los datos del viaje completado (M6) y del pago liquidado (M7) y genera el archivo PDF correspondiente.

        **Idempotencia (RNF-08):** Si ya existe un comprobante para el `tripId` enviado, la API responde `200 OK`
        con los datos del comprobante vigente en lugar de duplicar el archivo. Si es nuevo responde `201 Created`.

        **Autorización fiscal:** antes de generar el PDF se pide la autorización al autorizador fiscal externo
        (timeout de 2 s, circuit breaker). Si no responde o su circuito está abierto, responde
        `503 FISCAL_SERVICE_UNAVAILABLE` con `Retry-After` y no se persiste nada: el pedido se puede repetir.
        Si lo rechaza, responde `422 FISCAL_AUTHORIZATION_REJECTED`.

        **Pago verificado en M7 (2.2.0):** M7 es la fuente de verdad del pago. Antes de emitir se consulta
        `GET /metodo-pago/{viajeId}` de M7 (timeout de 2 s):
        - `autorizado`: se emite; `payment.status` queda `APROBADO` y `payment.method` se toma de M7
          (`efectivo`, `tarjeta`, `transferencia` → `EFECTIVO`, `TARJETA`, `TRANSFERENCIA`), aunque la entrada
          informe otro valor. Si M7 informa `total` y `moneda` (2.3.0), `fare.total` y `fare.currency` se toman
          de M7; el desglose de `fare` se conserva solo si suma ese total, y si no queda solo el total.
        - `pendiente` o sin pago registrado: `409 PAYMENT_PENDING` / `409 PAYMENT_NOT_FOUND` con `Retry-After`.
          No se persiste nada y el pedido se puede repetir cuando M7 autorice el pago.
        - `rechazado`: `422 PAYMENT_REJECTED`. Ese pago no genera comprobante.
        - M7 sin respuesta: `503 PAYMENTS_SERVICE_UNAVAILABLE` con `Retry-After`.
      operationId: createReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptRequest'
            example:
              tripId: "trip-2026-000123"
              customerUserId: 91
              driverUserId: 457
              customer:
                id: "cli-0091"
                fullName: "Lucia Fernandez"
                email: "lucia.fernandez@example.com"
                documentId: "38.442.019"
              driver:
                id: "cnd-0457"
                fullName: "Martin Rodriguez"
                vehicle:
                  type: "AUTO"
                  plate: "AB123CD"
                  model: "Toyota Etios 2021"
              trip:
                origin: "Av. Colon 1250, Cordoba"
                destination: "Aeropuerto Ambrosio Taravella"
                startedAt: "2026-08-28T13:05:00.000Z"
                finishedAt: "2026-08-28T13:36:00.000Z"
                distanceKm: 14.8
                durationMin: 31
              fare:
                currency: "ARS"
                baseFare: 1200
                distanceAmount: 5920
                timeAmount: 1550
                surcharges: 430
                discounts: 600
                total: 8500
              payment:
                method: "TARJETA"
                status: "APROBADO"
                authorizationCode: "AUTH-77321"
      responses:
        '201':
          description: Comprobante emitido exitosamente (nuevo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
        '200':
          description: Comprobante ya emitido previamente (idempotencia)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
        '400':
          description: Identificador de viaje o formato inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: |
            M7 todavía no autorizó el pago (`PAYMENT_PENDING`) o no tiene un pago registrado para el viaje
            (`PAYMENT_NOT_FOUND`). No se emite nada; el pedido se puede repetir después de `Retry-After`.
          headers:
            Retry-After:
              description: Segundos sugeridos antes de reintentar.
              schema:
                type: integer
                example: 5
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: PAYMENT_PENDING
                  message: El pago del viaje trip-2026-000123 esta pendiente de autorizacion en M7
                  path: /api/v1/receipts
                  timestamp: "2026-10-05T18:42:12.000Z"
        '422':
          description: |
            Error de validación en los campos del comprobante (`VALIDATION_ERROR`), M7 rechazó el pago
            (`PAYMENT_REJECTED`), o el autorizador fiscal rechazó el comprobante
            (`FISCAL_AUTHORIZATION_REJECTED`, con el motivo en `details`).
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
              examples:
                validacion:
                  summary: Datos inválidos
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: La solicitud contiene datos invalidos
                      path: /api/v1/receipts
                      timestamp: "2026-10-05T18:42:12.000Z"
                      details: ["customer.fullName es obligatorio"]
                rechazoFiscal:
                  summary: Rechazo del autorizador fiscal
                  value:
                    error:
                      code: FISCAL_AUTHORIZATION_REJECTED
                      message: El autorizador fiscal rechazo el comprobante
                      path: /api/v1/receipts
                      timestamp: "2026-10-05T18:42:12.000Z"
                      details: { code: AMOUNT_LIMIT_EXCEEDED, message: "El importe supera el maximo autorizable (10000000)" }
                pagoRechazado:
                  summary: Pago rechazado por M7
                  value:
                    error:
                      code: PAYMENT_REJECTED
                      message: M7 rechazo el pago del viaje trip-2026-000123
                      path: /api/v1/receipts
                      timestamp: "2026-10-05T18:42:12.000Z"
        '500':
          description: Error interno del servidor
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /receipts/{tripId}:
    get:
      tags:
        - Comprobantes (RF-8.3 / RF-8.4)
      summary: Consulta los metadatos y enlaces de descarga del comprobante
      description: >-
        Devuelve la información detallada del comprobante emitido, su historial de entregas y los enlaces de descarga.
        Requiere un Bearer de usuario validado por M1 y autorización sobre el comprobante.
      operationId: getReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          description: Identificador único del viaje
          schema:
            type: string
            example: "trip-2026-000123"
      responses:
        '200':
          description: Metadatos del comprobante encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptEnvelope'
        '400':
          description: Formato de tripId inválido
          content:
            application/json:
              schema:
              $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No existe un comprobante para el tripId indicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
  /receipts/resend:
    post:
      tags:
        - Comprobantes (RF-8.3 / RF-8.4)
      summary: Reenvío de comprobante de viaje por body (RF-8.4)
      description: |
        Vuelve a solicitar el envío del comprobante enviando el tripId en el cuerpo de la petición JSON.
      operationId: resendReceiptWithBody
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendRequest'
            example:
              tripId: "trip-2026-000123"
              channel: "EMAIL"
              destination: "lucia.fernandez@example.com"
      responses:
        '202':
          description: Reenvío aceptado y registrado en el historial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendEnvelope'
        '400':
          description: Formato de tripId inválido
          content:
            application/json:
              schema:
              $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Comprobante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El comprobante existe pero su archivo PDF no está disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Destino o canal de envío inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
  /receipts/{tripId}/resend:
    post:
      tags:
        - Comprobantes (RF-8.3 / RF-8.4)
      summary: Reenvío de comprobante de viaje (RF-8.4)
      description: |
        Vuelve a solicitar el envío del comprobante por correo electrónico, SMS o notificación push.
        Si no se indica destino en el cuerpo, utiliza el email registrado del cliente.
      operationId: resendReceipt
      security:
        - BearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          description: Identificador único del viaje
          schema:
            type: string
            example: "trip-2026-000123"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendRequest'
            example:
              channel: "EMAIL"
              destination: "lucia.fernandez@example.com"
      responses:
        '202':
          description: Reenvío aceptado y registrado en el historial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResendEnvelope'
        '400':
          description: Formato de tripId inválido
          content:
            application/json:
              schema:
              $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Comprobante no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: El comprobante existe pero su archivo PDF no está disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Destino o canal de envío inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '503':
          $ref: '#/components/responses/DependencyUnavailable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer del usuario validado por M1 mediante su endpoint de identidad.
  responses:
    Unauthorized:
      description: Falta el Bearer o M1 no lo reconoce.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no es dueño del comprobante ni operador autorizado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    DependencyUnavailable:
      description: |
        Una dependencia no está disponible: la base de datos (`DATABASE_UNAVAILABLE`), el autorizador
        fiscal (`FISCAL_SERVICE_UNAVAILABLE`, sin respuesta o con el circuito abierto), la API de pagos
        de M7 (`PAYMENTS_SERVICE_UNAVAILABLE`), o el generador de enlaces (`DOWNLOAD_LINKS_UNAVAILABLE`).
        El pedido se puede repetir después del tiempo indicado en `Retry-After`.
      headers:
        Retry-After:
          description: Segundos sugeridos antes de reintentar.
          schema:
            type: integer
            example: 5
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FISCAL_SERVICE_UNAVAILABLE
              message: El autorizador fiscal no esta disponible (circuito abierto)
              path: /api/v1/receipts
              timestamp: "2026-10-05T18:42:12.000Z"
  schemas:
...
    ReceiptRequest:
      type: object
      required:
        - tripId
        - customer
        - driver
        - trip
        - fare
        - payment
      properties:
        tripId:
          type: string
          pattern: '^[A-Za-z0-9_-]{1,64}$'
          example: "trip-2026-000123"
        customerUserId:
          type: integer
          format: int64
          minimum: 1
          description: Identificador canónico numérico del cliente en M1. Se conserva para autorizar consultas y reenvíos de RF-8.4.
          example: 91
        driverUserId:
          type: integer
          format: int64
          minimum: 1
          description: Identificador canónico numérico del conductor en M1. Se conserva para autorizar consultas y reenvíos de RF-8.4.
          example: 457
        customer:
          $ref: '#/components/schemas/Customer'
        driver:
          $ref: '#/components/schemas/Driver'
        trip:
          $ref: '#/components/schemas/TripDetails'
        fare:
          $ref: '#/components/schemas/FareDetails'
        payment:
          $ref: '#/components/schemas/PaymentDetails'
    Customer:
      type: object
      required:
        - id
        - fullName
      properties:
        id:
          type: string
          example: "cli-0091"
        fullName:
          type: string
          example: "Lucia Fernandez"
        email:
          type: string
          format: email
          example: "lucia.fernandez@example.com"
        phone:
          type: string
          example: "+54 9 351 555-0199"
        documentId:
          type: string
          example: "38.442.019"
    Driver:
      type: object
      required:
        - id
        - fullName
        - vehicle
      properties:
        id:
          type: string
          example: "cnd-0457"
        fullName:
          type: string
          example: "Martin Rodriguez"
        vehicle:
          $ref: '#/components/schemas/Vehicle'
    Vehicle:
      type: object
      required:
        - type
        - plate
      properties:
        type:
          type: string
          enum: [AUTO, MOTO]
          example: "AUTO"
        plate:
          type: string
          example: "AB123CD"
        model:
          type: string
          example: "Toyota Etios 2021"
    TripDetails:
      type: object
      required:
        - origin
        - destination
        - startedAt
        - finishedAt
        - distanceKm
      properties:
        origin:
          type: string
          example: "Av. Colon 1250, Cordoba"
        destination:
          type: string
          example: "Aeropuerto Ambrosio Taravella"
        startedAt:
          type: string
          format: date-time
          example: "2026-08-28T13:05:00.000Z"
        finishedAt:
          type: string
          format: date-time
          example: "2026-08-28T13:36:00.000Z"
        distanceKm:
          type: number
          minimum: 0
          example: 14.8
        durationMin:
          type: integer
          minimum: 0
          example: 31
    FareDetails:
      type: object
      required:
        - total
      properties:
        currency:
          type: string
          default: "ARS"
          example: "ARS"
        baseFare:
          type: number
          minimum: 0
          example: 1200
        distanceAmount:
          type: number
          minimum: 0
          example: 5920
        timeAmount:
          type: number
          minimum: 0
          example: 1550
        surcharges:
          type: number
          minimum: 0
          example: 430
        discounts:
          type: number
          minimum: 0
          example: 600
        total:
          type: number
          minimum: 0
          example: 8500
    PaymentDetails:
      type: object
      required:
        - method
        - status
      properties:
        method:
          type: string
          enum: [EFECTIVO, TARJETA, TRANSFERENCIA, BILLETERA]
          description: |
            Al emitir se toma de M7. `TRANSFERENCIA` corresponde a `transferencia` de M7; `BILLETERA` no existe
            en M7 y se conserva por compatibilidad.
          example: "TARJETA"
        status:
          type: string
          enum: [APROBADO, PENDIENTE, RECHAZADO]
          example: "APROBADO"
        authorizationCode:
          type: string
          example: "AUTH-77321"
    ReceiptEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Receipt'
    Receipt:
      type: object
      properties:
        receiptId:
          type: string
          format: uuid
          example: "cee84adf-d510-495c-b2be-6def62b11409"
        receiptNumber:
          type: string
          description: Numero legible del comprobante, con el formato CMP-<anio>-<referencia>.
          example: "CMP-2026-CEE84ADFD5"
        tripId:
          type: string
          example: "trip-2026-000123"
        issuedAt:
          type: string
          format: date-time
          example: "2026-08-28T13:37:00.000Z"
        customer:
          $ref: '#/components/schemas/Customer'
        driver:
          $ref: '#/components/schemas/Driver'
        trip:
          $ref: '#/components/schemas/TripDetails'
        fare:
          $ref: '#/components/schemas/FareDetails'
        payment:
          $ref: '#/components/schemas/PaymentDetails'
        fiscal:
          $ref: '#/components/schemas/FiscalAuthorization'
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryRecord'
        pdf:
          type: object
          properties:
            downloadUrl:
              type: string
              example: "http://localhost:3008/api/v1/receipts/trip-2026-000123/pdf"
    FiscalAuthorization:
      type: object
      description: |
        Autorización otorgada por el autorizador fiscal externo (simulado). Ausente solo en comprobantes
        emitidos antes de la versión 2.1.0.
      required: [authorizationCode, expiresOn, authorizedAt]
      properties:
        authorizationCode:
          type: string
          pattern: '^\d{14}$'
          example: "71234567890123"
        expiresOn:
          type: string
          format: date
          example: "2026-10-15"
        authorizedAt:
          type: string
          format: date-time
          example: "2026-10-05T18:42:12.000Z"
    DeliveryRecord:
      type: object
      description: |
        Constancia de una entrega del comprobante. En AE1 el envio se simula: se
        registra el pedido y se devuelve el enlace de descarga.
      required:
        - channel
        - destination
        - sentAt
      properties:
        channel:
          type: string
          enum: [EMAIL, SMS, PUSH]
          example: "EMAIL"
        destination:
          type: string
          description: Destino al que se solicito la entrega. En los logs del servicio se enmascara.
          example: "lucia.fernandez@example.com"
        sentAt:
          type: string
          format: date-time
          example: "2026-08-28T10:50:40.449Z"
    ResendRequest:
      type: object
      properties:
        tripId:
          type: string
          description: "Opcional. Identificador del viaje (si se envía por el cuerpo en lugar de la URL)."
          example: "trip-2026-000123"
        channel:
          type: string
          enum: [EMAIL, SMS, PUSH]
          default: "EMAIL"
          example: "EMAIL"
        destination:
          type: string
          example: "lucia.fernandez@example.com"
    ResendEnvelope:
      type: object
      properties:
        data:
          allOf:
            - $ref: '#/components/schemas/Receipt'
            - type: object
              properties:
                lastDelivery:
                  allOf:
                    - $ref: '#/components/schemas/DeliveryRecord'
                    - type: object
                      properties:
                        url:
                          type: string
                          example: "http://localhost:3008/api/v1/receipts/downloads/random-token-here"
                        expiresAt:
                          type: string
                          format: date-time
                          example: "2026-08-28T14:37:00.000Z"
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - path
            - timestamp
          properties:
            code:
              type: string
              example: "RECEIPT_NOT_FOUND"
            message:
              type: string
              example: "No existe un comprobante emitido para el viaje trip-2026-000123"
            path:
              type: string
              example: "/api/v1/receipts/trip-2026-000123"
            timestamp:
              type: string
              format: date-time
              example: "2026-08-28T10:55:36.653Z"
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - path
            - timestamp
            - details
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "La solicitud contiene datos invalidos"
            path:
              type: string
              example: "/api/v1/receipts"
            timestamp:
              type: string
              format: date-time
              example: "2026-08-28T10:55:36.731Z"
            details:
              type: array
              description: Lista de los campos que no cumplen el contrato, uno por mensaje.
              items:
                type: string
              example:
                - "customer.fullName es obligatorio"
                - "fare.total (100) no coincide con el desglose informado (50)"