# modulo-8/services/qr/src/http/api-error.ts
export interface ErrorDetail {
  field: string;
  reason: string;
}

export interface ErrorResponse {
  error: {
    code: string;
    message: string;
    details?: ErrorDetail[];
  };
}

export class ApiError extends Error {
  constructor(
    public readonly status: number,
    public readonly code: string,
    message: string,
    public readonly details?: ErrorDetail[],
    // Headers HTTP que acompañan la respuesta de error (por ejemplo, Retry-After).
    public readonly headers?: Readonly<Record<string, string>>,
  ) {
    super(message);
    this.name = "ApiError";
  }
}